moved all auth endpoints to routes/auth, moved auth to last to prevent auth will testing
